API/requirements.txt
flask
Pillow
gunicorn
requests
mysql-connector-python